Remove styling for utility nav since it's dependent on its parent component
