--- src/components/navigation/utility-nav/UtilityNavStyles.js
@@ -1,1 +1,0 @@
-export const root = 'su-bg-digital-red';
