+++ src/components/storyblok/utilityNav.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import SbEditable from 'storyblok-react';
-import UtilityNav from '../navigation/utility-nav/UtilityNav';
+import UtilityNav from '../navigation/UtilityNav';
 
 export const SBUtilityNav = ({ blok }) => {
   const { menuItems } = blok;
